Some Changes in REST-API & in UI

+++ JSS/REST-API/JSS-Employee/src/main/java/com/jss/employee/dto/Employee.java
@@ -16,7 +16,7 @@ public class Employee {
 	private int experience;
 	private int salary;
 	private int noticePeriod;
-	private int contact;
+	private long contact;
 	private String email;
 	private String status;
 	private String role;
@@ -27,7 +27,7 @@ public class Employee {
 	}
 	
 	public Employee(int id, String userName, String password, String orgnizationName, String address, String jobTitle,
-			String location, String description, int experience, int salary, int noticePeriod, int contact,
+			String location, String description, int experience, int salary, int noticePeriod, long contact,
 			String email, String status, String role) {
 	
 		this.id = id;
@@ -112,10 +112,10 @@ public class Employee {
 	public void setNoticePeriod(int noticePeriod) {
 		this.noticePeriod = noticePeriod;
 	}
-	public int getContact() {
+	public long getContact() {
 		return contact;
 	}
-	public void setContact(int contact) {
+	public void setContact(long contact) {
 		this.contact = contact;
 	}
 	public String getEmail() {
